Now using attributes to define operation and argument metadata. Operation metadata is now set using the OperationDefinitionAttribute, rather than the constructor.
Arguments work differently now. Instead of using an IArgument instance to specify an argument on an operation, you now specify the base type (int, string, bool), and mark it with the OperationArgumentAttribute, supplying the ID, friendly name, description, etc. The OperationController will take care of retrieving the argument value, if there is one, and will use reflection to write the value to the argument field on the Operation instance.